(ios) Minor refactoring
